# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Got Journey Thailand API",
    "description": "Travel Booking Platform API - Hotels, Cars, and Package Bookings",
    "version": "1.0.0",
    "contact": {
      "name": "Got Journey Thailand",
      "email": "support@gotjourneythailand.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development Server"
    },
    {
      "url": "https://api.gotjourneythailand.com",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/api/hotels/{id}": {
      "get": {
        "tags": ["Hotels"],
        "summary": "Get hotel by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "lang", "in": "query", "schema": { "type": "string", "enum": ["th", "en"] } }
        ],
        "responses": {
          "200": { "description": "Successful response" },
          "404": { "description": "Hotel not found" }
        }
      },
      "put": {
        "tags": ["Hotels"],
        "summary": "Update hotel (Admin only)",
        "security": [{ "AdminAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/HotelInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Hotel updated" },
          "404": { "description": "Hotel not found" }
        }
      },
      "delete": {
        "tags": ["Hotels"],
        "summary": "Delete hotel (Admin only)",
        "security": [{ "AdminAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Hotel deleted" },
          "404": { "description": "Hotel not found" }
        }
      }
    },
    "/api/cars": {
      "get": {
        "tags": ["Cars"],
        "summary": "Get all cars",
        "parameters": [
          { "name": "active", "in": "query", "schema": { "type": "boolean" } },
          { "name": "lang", "in": "query", "schema": { "type": "string", "enum": ["th", "en"] } }
        ],
        "responses": {
          "200": { "description": "Successful response" }
        }
      },
      "post": {
        "tags": ["Cars"],
        "summary": "Create a new car (Admin only)",
        "security": [{ "AdminAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CarInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Car created" }
        }
      }
    },
...
    "/api/checkout": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create Stripe checkout session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["booking_code"],
                "properties": {
                  "booking_code": { "type": "string" },
                  "booking_id": { "type": "string" },
                  "currency": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": { "type": "string" },
                    "session_id": { "type": "string" },
                    "booking_code": { "type": "string" },
                    "amount": { "type": "number" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_token"
      },
      "UserAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_token"
      }
    },
    "schemas": {
...
      "HotelInput": {
        "type": "object",
        "required": ["name_th", "name_en"],
        "properties": {
          "name_th": { "type": "string" },
          "name_en": { "type": "string" },
          "description_th": { "type": "string" },
          "description_en": { "type": "string" },
          "location_th": { "type": "string" },
          "location_en": { "type": "string" },
          "star_rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "base_price_per_night": { "type": "number" },
          "max_guests": { "type": "integer" },
          "room_type_th": { "type": "string" },
          "room_type_en": { "type": "string" },
          "amenities_th": { "type": "array", "items": { "type": "string" } },
          "amenities_en": { "type": "array", "items": { "type": "string" } },
          "images": { "type": "array", "items": { "type": "string" } }
        }
      },
      "CarInput": {
        "type": "object",
        "required": ["name_th", "name_en"],
        "properties": {
          "name_th": { "type": "string" },
          "name_en": { "type": "string" },
          "description_th": { "type": "string" },
          "description_en": { "type": "string" },
          "car_type_th": { "type": "string" },
          "car_type_en": { "type": "string" },
          "max_passengers": { "type": "integer" },
          "base_price_per_day": { "type": "number" },
          "includes_th": { "type": "array", "items": { "type": "string" } },
          "includes_en": { "type": "array", "items": { "type": "string" } },
          "images": { "type": "array", "items": { "type": "string" } }
        }
      },
...
    }
  }
}